fix: revert folder structure refactor
