+++ .gitignore
@@ -35,6 +35,3 @@ yarn-error.log*
 # typescript
 *.tsbuildinfo
 next-env.d.ts
-
-# .windsurfrules
-.windsurfrules
