--- src/features/post/actions/index.ts
@@ -1,5 +1,0 @@
-export * from './create';
-export * from './get';
-export * from './like';
-export * from './comment';
-export * from './delete';
