+++ src/types/index.ts
@@ -1,3 +1,3 @@
-import { getPostById } from "@/features/post/actions";
+import { getPostById } from "@/actions/post.action";
 
 export type Post = Awaited<ReturnType<typeof getPostById>>;
